+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.1.0
+- Added compatibility (soft dependency) with Coroner and custom death messages
+- Added Adrenaline rush as a third effect to occur
+- Added more config options
+- Added ability for adrenaline to cause a tremble at the end
+
 ## v1.0.2
 - Updated `README.md` for clarity
 - Added `CHANGELOG.md`